Class to Hooks, moved to own component file
