+++ src/components/search/Search.scss
@@ -1,5 +1,5 @@
-@import '../styles/variables';
-@import '../styles/mixins.scss';
+@import '../../styles/variables';
+@import '../../styles/mixins.scss';
 
 .va-api-search-wrapper {
   button[type=submit] {
